+++ lib/mapi/oxcical.cpp
@@ -753,7 +753,7 @@ static BOOL oxcical_parse_recipients(const ical_component &main_ev,
 		tmp_int32 = role_to_rcpttype(prole, cutype);
 		if (pproplist->set(PR_RECIPIENT_TYPE, &tmp_int32) != 0)
 			return FALSE;
-		tmp_int32 = dtypx == DT_DISTLIST ? MAPI_DISTLIST : MAPI_MAILUSER;
+		tmp_int32 = static_cast<uint32_t>(dtypx == DT_DISTLIST ? MAPI_DISTLIST : MAPI_MAILUSER);
 		if (pproplist->set(PR_OBJECT_TYPE, &tmp_int32) != 0)
 			return FALSE;
 		tmp_int32 = static_cast<uint32_t>(dtypx);
