+++ tools/eml2mt.cpp
@@ -290,9 +290,9 @@ int main(int argc, const char **argv) try
 			fprintf(stderr, "E-2020: ENOMEM\n");
 			return EXIT_FAILURE;
 		}
-		if (ep.p_uint32(MAPI_MESSAGE) != EXT_ERR_SUCCESS ||
+		if (ep.p_uint32(static_cast<uint32_t>(MAPI_MESSAGE)) != EXT_ERR_SUCCESS ||
 		    ep.p_uint32(i + 1) != EXT_ERR_SUCCESS ||
-		    ep.p_uint32(parent.type) != EXT_ERR_SUCCESS ||
+		    ep.p_uint32(static_cast<uint32_t>(parent.type)) != EXT_ERR_SUCCESS ||
 		    ep.p_uint64(parent.folder_id) != EXT_ERR_SUCCESS ||
 		    ep.p_msgctnt(*msgs[i]) != EXT_ERR_SUCCESS) {
 			fprintf(stderr, "E-2021\n");
