+++ tools/oxm2mt.cpp
@@ -579,9 +579,9 @@ static errno_t do_file(const char *filename) try
 		fprintf(stderr, "E-2020: ENOMEM\n");
 		return EXIT_FAILURE;
 	}
-	if (ep.p_uint32(MAPI_MESSAGE) != EXT_ERR_SUCCESS ||
+	if (ep.p_uint32(static_cast<uint32_t>(MAPI_MESSAGE)) != EXT_ERR_SUCCESS ||
 	    ep.p_uint32(1) != EXT_ERR_SUCCESS ||
-	    ep.p_uint32(parent.type) != EXT_ERR_SUCCESS ||
+	    ep.p_uint32(static_cast<uint32_t>(parent.type)) != EXT_ERR_SUCCESS ||
 	    ep.p_uint64(parent.folder_id) != EXT_ERR_SUCCESS ||
 	    ep.p_msgctnt(*ctnt) != EXT_ERR_SUCCESS) {
 		fprintf(stderr, "E-2021\n");
